requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-multipart==0.0.20
requests==2.32.4
httpx==0.28.1
Pillow==11.3.0
Werkzeug==3.1.3
pydantic==2.11.7
pydantic-settings==2.10.1
moviepy==2.2.1
edge-tts==7.2.3
langdetect==1.0.9
piper-tts==1.3.0
tqdm==4.67.1
nltk==3.9.2
numpy==2.3.2
torch==2.8.0
diffusers==0.35.1
accelerate==1.10.0
transformers==4.56.1
sentence-transformers==5.1.0
openai-whisper==20250625
pytest==8.4.1
